refactor: remove dead re-exports, thin wrappers, and consolidate modules
Phase 1: Remove unused pub use re-exports from lib.rs, schema/mod.rs,
narrowing public API surface.

Phase 2: Remove thin delegation methods from DbOperations (get_item,
delete_item, batch_store_in_namespace). Callers now use typed stores
directly (e.g., db_ops.atoms_store().get_item()).

Phase 3: Merge 5 single-use ingestion submodules into their callers:
ai_schema_response→ai_helpers, prompts→ai_helpers,
date_handling→key_extraction, cost_estimation→smart_folder,
multipart_parser→file_upload.

Phase 4: Deduplicate SuccessResponse (defined 3 times) into a single
definition in handlers/response.rs.

Net: -974 lines removed, 5 files deleted, cleaner module boundaries.
